# optimize slow tests