<!--
Please make sure you've read and understood our contributing guidelines;
https://github.com/docker/docker/blob/master/CONTRIBUTING.md

** Make sure all your commits include a signature generated with `git commit -s` **

For additional information on our contributing process, read our contributing
guide https://docs.docker.com/opensource/code/

If this is a bug fix, make sure your description includes "fixes #xxxx", or
"closes #xxxx"

Please provide the following information:
-->

**- What I did**
Optimize 2 slow tests.

Before:
```
PASS: docker_cli_run_test.go:2751: DockerSuite.TestRunContainerWithReadonlyRootfs 16.561s
PASS: docker_cli_run_unix_test.go:910: DockerSuite.TestRunApparmorProcDirectory 19.075s
```

After:
```
PASS: docker_cli_run_test.go:2751: DockerSuite.TestRunContainerWithReadonlyRootfs	2.261s
PASS: docker_cli_run_unix_test.go:910: DockerSuite.TestRunApparmorProcDirectory	2.250s
```
**- How I did it**

1. Merge multiple `docker run` command into one for ` TestRunContainerWithReadonlyRootfs `
2. Change the image from `debian:jessie` to `busybox` for `TestRunApparmorProcDirectory`

**- How to verify it**
`make test-integration-cli`

**- A picture of a cute animal (not mandatory but encouraged)**
![1c767db5c9ea15ce0f140dbcb5003af33b87b25e](https://cloud.githubusercontent.com/assets/1496652/13722189/b88ff2c8-e877-11e5-83ca-4465da2e0ab8.jpg)


Signed-off-by: Shijiang Wei <mountkin@gmail.com>