@cyphar 2016-03-12
I'm wondering whether this invalidates the test (since the old code was trying to make sure that *all* of them fail, but the current code only makes sure that at least *one* of them fails). For example, what happens if we have a regression with `/sys` becoming `rw` even with `--read-only`?

Maybe if you made it a `for` loop *in the shell script* then you could check the output for each file -- that should increase the speed without removing the validity of the test. Something like:

```
docker run --read-only sh -c 'for i in <list of files> { touch $i; echo "$? $i"; }'
```

Or similar (you could even use the error output from `touch`.

:+1: on switching from `debian:jessie` to `busybox`. Maybe we should look through the tests for more occurrences of that.