@mountkin 2016-03-12
> I'm wondering whether this invalidates the test (since the old code was trying to make sure that all of them fail, but the current code only makes sure that at least one of them fails). For example, what happens if we have a regression with /sys becoming rw even with --read-only?

Oh it's really a problem. 
Thanks for pointing out this @cyphar :+1: 
I'll update the code accordingly.